pyproject.toml
[project]
name = "yt-dlp-dj"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "yt-dlp",
    "mutagen",
    "typer",
    "pyyaml",
    "pydantic",
]

[project.scripts]
yt-dlp-dj = "server.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.mypy]
strict = true
packages = ["server"]

[[tool.mypy.overrides]]
module = "mutagen.*"
follow_imports = "skip"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501", "SIM117"]  # no line-length enforcement; nested with is fine

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
    "types-pyyaml>=6.0.12.20250915",
    "httpx>=0.27",
    "pytest-asyncio>=0.24",
]
